rl_agent/requirements.txt
gymnasium>=0.29
stable-baselines3>=2.0
numpy
torch>=2.0
PyYAML
wandb
pytest
